Add control bar - music on/off and texture changing Split script.js into script.js(universal threejs setup) and marsglobe.js Add favicon.svg
